add status helper text.
