+++ core/commands/storage/upload/status_cmd.go
