+++ core/commands/storage/upload/upload_cmd_renter_sign_guard_shard_contracts.go
